moves store/contracts state to types file
