+++ src/components/web3/helpers/getAdapterAddress.ts
@@ -1,6 +1,6 @@
 import Web3 from 'web3';
 
-import {ContractsStateEntry} from '../../../store/types';
+import {ContractsStateEntry} from '../../../store/contracts/types';
 import {ContractAdapterNames} from '../types';
 
 export async function getAdapterAddress(
